refactor: separate commands and controllers
