+++ package.json
@@ -34,7 +34,7 @@
     "onCommand:fileutils.newFolder",
     "onCommand:fileutils.newFolderAtRoot"
   ],
-  "main": "./out/src/extension",
+  "main": "./out/src",
   "contributes": {
     "commands": [
       {
