+++ src/command/RemoveFileCommand.ts
@@ -1,0 +1,10 @@
+import { RemoveFileController } from '../controller';
+
+export const controller = new RemoveFileController();
+
+export function removeFile() {
+
+    return controller.showDialog()
+        .then((fileItem) => controller.remove(fileItem))
+        .then(() => controller.closeCurrentFileEditor());
+}
