+++ src/command/index.ts
@@ -1,0 +1,5 @@
+export { newFile, newFileAtRoot, newFolder, newFolderAtRoot } from './NewFileCommand';
+export { moveFile } from './MoveFileCommand';
+export { renameFile } from './RenameFileCommand';
+export { duplicateFile } from './DuplicateFileCommand';
+export { removeFile } from './RemoveFileCommand';
