+++ src/controller/index.ts
@@ -1,0 +1,4 @@
+export { DuplicateFileController } from './DuplicateFileController';
+export { MoveFileController } from './MoveFileController';
+export { NewFileController } from './NewFileController';
+export { RemoveFileController } from './RemoveFileController';
