+++ test/command/DuplicateFileTest.ts
@@ -1,31 +1,17 @@
 import * as retry from 'bluebird-retry';
-import {
-    expect,
-    use as chaiUse
-} from 'chai';
+import { expect, use as chaiUse } from 'chai';
 import * as fs from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
 import * as sinon from 'sinon';
 import * as sinonChai from 'sinon-chai';
-
-import {
-    commands,
-    TextEditor,
-    Uri,
-    window,
-    workspace
-} from 'vscode';
-
-import {
-    controller,
-    moveFile
-} from '../../src/extension/commands';
+import { commands, TextEditor, Uri, window, workspace } from 'vscode';
+import { controller, duplicateFile } from '../../src/command/DuplicateFileCommand';
 
 chaiUse(sinonChai);
 
 const rootDir = path.resolve(__dirname, '..', '..', '..');
-const tmpDir = path.resolve(os.tmpdir(), 'vscode-fileutils-test--move-file');
+const tmpDir = path.resolve(os.tmpdir(), 'vscode-fileutils-test--duplicate-file');
 
 const fixtureFile1 = path.resolve(rootDir, 'test', 'fixtures', 'file-1.rb');
 const fixtureFile2 = path.resolve(rootDir, 'test', 'fixtures', 'file-2.rb');
@@ -35,7 +21,7 @@ const editorFile2 = path.resolve(tmpDir, 'file-2.rb');
 
 const targetFile = path.resolve(`${editorFile1}.tmp`);
 
-describe('moveFile', () => {
+describe('duplicateFile', () => {
 
     beforeEach(() => Promise.all([
         fs.remove(tmpDir),
@@ -87,8 +73,8 @@ describe('moveFile', () => {
 
             it('prompts for file destination', () => {
 
-                return moveFile().then(() => {
-                    const prompt = 'New Location';
+                return duplicateFile().then(() => {
+                    const prompt = 'Duplicate As';
                     const value = editorFile1;
                     const valueSelection = [value.length - 9, value.length - 3];
                     expect(window.showInputBox).to.have.been.calledWithExactly({ prompt, value, valueSelection });
@@ -98,7 +84,7 @@ describe('moveFile', () => {
 
             it('moves current file to destination', () => {
 
-                return moveFile().then(() => {
+                return duplicateFile().then(() => {
                     const message = `${targetFile} does not exist`;
                     // tslint:disable-next-line:no-unused-expression
                     expect(fs.existsSync(targetFile), message).to.be.true;
@@ -116,7 +102,7 @@ describe('moveFile', () => {
 
                 it('creates nested directories', () => {
 
-                    return moveFile().then((textEditor: TextEditor) => {
+                    return duplicateFile().then((textEditor: TextEditor) => {
                         const dirname = path.dirname(textEditor.document.fileName);
                         const directories: string[] = dirname.split(path.sep);
 
@@ -130,7 +116,7 @@ describe('moveFile', () => {
 
             it('opens target file as active editor', () => {
 
-                return moveFile().then(() => {
+                return duplicateFile().then(() => {
                     const activeEditor: TextEditor = window.activeTextEditor;
                     expect(activeEditor.document.fileName).to.equal(targetFile);
                 });
@@ -166,16 +152,16 @@ describe('moveFile', () => {
                     const action = 'Overwrite';
                     const options = { modal: true };
 
-                    return moveFile().then(() => {
+                    return duplicateFile().then(() => {
                         expect(window.showInformationMessage).to.have.been.calledWith(message, options, action);
                     });
                 });
 
-                describe('responding with Overwrite', () => {
+                describe(`responding with 'Overwrite'`, () => {
 
-                    it('overwrites the existing file', () => {
+                    it('overwrites the existig file', () => {
 
-                        return moveFile().then(() => {
+                        return duplicateFile().then(() => {
                             const fileContent = fs.readFileSync(targetFile).toString();
                             expect(fileContent).to.equal('class FileOne; end');
                         });
@@ -183,7 +169,7 @@ describe('moveFile', () => {
 
                 });
 
-                describe('responding with Cancel', () => {
+                describe(`responding with 'Cancel'`, () => {
 
                     beforeEach(() => {
                         const stub: any = window.showInformationMessage;
@@ -193,7 +179,7 @@ describe('moveFile', () => {
 
                     it('leaves existing file untouched', () => {
 
-                        return moveFile().then(() => {
+                        return duplicateFile().then(() => {
                             const fileContent = fs.readFileSync(targetFile).toString();
                             expect(fileContent).to.equal('class FileTwo; end');
                         });
@@ -231,7 +217,7 @@ describe('moveFile', () => {
 
             it('ignores the command call', () => {
 
-                return moveFile().catch(() => {
+                return duplicateFile().catch(() => {
                     // tslint:disable-next-line:no-unused-expression
                     expect(window.showInputBox).to.have.not.been.called;
                 });
@@ -255,8 +241,8 @@ describe('moveFile', () => {
 
         it('prompts for file destination', () => {
 
-            return moveFile(Uri.file(editorFile1)).then(() => {
-                const prompt = 'New Location';
+            return duplicateFile(Uri.file(editorFile1)).then(() => {
+                const prompt = 'Duplicate As';
                 const value = editorFile1;
                 const valueSelection = [value.length - 9, value.length - 3];
                 expect(window.showInputBox).to.have.been.calledWithExactly({ prompt, value, valueSelection });
@@ -264,9 +250,9 @@ describe('moveFile', () => {
 
         });
 
-        it('moves current file to destination', () => {
+        it('duplicates current file to destination', () => {
 
-            return moveFile(Uri.file(editorFile1)).then(() => {
+            return duplicateFile(Uri.file(editorFile1)).then(() => {
                 const message = `${targetFile} does not exist`;
                 // tslint:disable-next-line:no-unused-expression
                 expect(fs.existsSync(targetFile), message).to.be.true;
@@ -275,7 +261,7 @@ describe('moveFile', () => {
 
         it('opens target file as active editor', () => {
 
-            return moveFile(Uri.file(editorFile1)).then(() => {
+            return duplicateFile(Uri.file(editorFile1)).then(() => {
                 const activeEditor: TextEditor = window.activeTextEditor;
                 expect(activeEditor.document.fileName).to.equal(targetFile);
             });
@@ -286,7 +272,7 @@ describe('moveFile', () => {
     describe('error handling', () => {
 
         beforeEach(() => {
-            sinon.stub(controller, 'showMoveFileDialog').returns(Promise.reject('must fail'));
+            sinon.stub(controller, 'showDialog').returns(Promise.reject('must fail'));
             sinon.stub(window, 'showErrorMessage');
             return Promise.resolve();
         });
@@ -294,7 +280,7 @@ describe('moveFile', () => {
         afterEach(() => {
 
             const restoreShowMoveFileDialog = () => {
-                const stub: any = controller.showMoveFileDialog;
+                const stub: any = controller.showDialog;
                 return Promise.resolve(stub.restore());
             };
 
@@ -311,7 +297,7 @@ describe('moveFile', () => {
 
         it('shows an error message', () => {
 
-            return moveFile().catch(() => {
+            return duplicateFile().catch((err) => {
                 expect(window.showErrorMessage).to.have.been.calledWithExactly('must fail');
             });
         });
